posting과 관련된 유틸 따로 분리하기
