+++ src/utils/postUtils.ts
@@ -13,7 +13,6 @@ import {
   QueryDocumentSnapshot
 } from 'firebase/firestore';
 
-import { Posting } from '@/types/Posting';
 import { firestore } from '../firebase';
 import { Post, PostVisibility } from '../types/Post';
 
@@ -27,37 +26,6 @@ export function mapDocumentToPost(snapshot: DocumentSnapshot | QueryDocumentSnap
     ...data,
     id: snapshot.id // 스냅샷 ID를 항상 사용
   };
-}
-
-/**
- * Firebase 문서를 Posting 객체로 변환하는 유틸리티 함수
- */
-export function mapDocumentToPosting(doc: QueryDocumentSnapshot): Posting {
-  const data = doc.data() as Posting;
-  // Timestamp 보장
-  data.createdAt = ensureTimestamp(data.createdAt);
-  
-  // post.id가 없을 경우 Firebase 문서 ID 할당
-  if (!data.post.id) {
-      data.post.id = doc.id;
-  }
-  
-  return data;
-}
-
-
-function ensureTimestamp(value: any): Timestamp {
-  if (value instanceof Timestamp) {
-      return value;
-  }
-
-  if (value && 
-      typeof value.seconds === "number" && 
-      typeof value.nanoseconds === "number") {
-      return new Timestamp(value.seconds, value.nanoseconds);
-  }
-
-  return Timestamp.now();
 }
 
 export const fetchPost = async (boardId: string, postId: string): Promise<Post | null> => {
